fix(session): address PR #90 round 2 — 3 defects + repo conventions
P2 — plain-text search hit the FTS5 parser. `SessionSearchParams::query` is
documented as plain text, but was bound straight to `MATCH`, so ordinary input
(`C++`, `foo-bar`, `file.rs`, a stray quote) returned a syntax or `no such
column` error instead of results. Each whitespace-separated term is now emitted
as a quoted FTS5 string literal (quotes escaped by doubling) joined with AND,
so punctuation is data and multi-term queries stay conjunctive.

P2 — an upsert could strand a claimed task. A claim only means anything while
the task is `in_progress`, but editing a claimed task through the upsert left
`claimed_by_member_id`/`claim_token` set on a `todo` row: a new claim saw
AlreadyClaimed, completion saw NotClaimed, and release/shutdown skipped it
because they only match `in_progress`. The claim is now cleared whenever the
new status is not `in_progress`, and preserved when it is, so an unrelated edit
cannot steal a live claim.

P2 — a failed completion gate discarded submitted evidence. Evidence
accumulates across attempts, so dropping it punished the caller for an
unrelated gate failure: after fixing a dependency, a retry that did not resend
the same links failed `require_evidence` on evidence already submitted. The
failed-gate update now persists the merged evidence alongside the verdict.

P1 — test layout. AGENTS.md requires module-local tests in a dedicated
`test.rs` per module directory; this module had `ops_tests.rs` plus inline
`mod tests` blocks in four files, which would have established a conflicting
convention in a large new feature. Consolidated into `session/test.rs` and
`session/run_ledger/test.rs`, with the helpers they reach widened to
`pub(super)`. All 62 session tests survive the move.

P1 — missing module README. AGENTS.md requires one for complex modules.
Added `src/session/README.md` covering the schema, the FTS behaviour, the
coordination guarantees, and the operational constraints — each of the
non-obvious rules there is pinned by a named test.

1424 tests pass, clippy clean under -D warnings, fmt clean.
