+++ src/session/types.rs
@@ -105,44 +105,3 @@ pub struct SessionSearchResult {
     pub sessions: Vec<SessionRecord>,
     pub total: u64,
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn session_status_roundtrip() {
-        for status in [
-            SessionStatus::Running,
-            SessionStatus::Completed,
-            SessionStatus::Failed,
-            SessionStatus::Interrupted,
-        ] {
-            assert_eq!(SessionStatus::parse(status.as_str()), status);
-        }
-    }
-
-    #[test]
-    fn session_status_parse_unknown_defaults_to_running() {
-        assert_eq!(SessionStatus::parse("bogus"), SessionStatus::Running);
-        assert_eq!(SessionStatus::parse(""), SessionStatus::Running);
-    }
-
-    #[test]
-    fn session_status_serde_roundtrip() {
-        let status = SessionStatus::Completed;
-        let json = serde_json::to_string(&status).unwrap();
-        assert_eq!(json, "\"completed\"");
-        let parsed: SessionStatus = serde_json::from_str(&json).unwrap();
-        assert_eq!(parsed, status);
-    }
-
-    #[test]
-    fn session_search_params_defaults() {
-        let params = SessionSearchParams::default();
-        assert!(params.query.is_none());
-        assert!(params.agent_id.is_none());
-        assert!(params.limit.is_none());
-        assert!(params.offset.is_none());
-    }
-}
